+++ aim/web/ui/src/pages/ImagesExplore/components/SelectForm/SelectForm.tsx
@@ -8,8 +8,8 @@ import {
 } from '@material-ui/icons';
 
 import { Icon, Badge, Button } from 'components/kit';
-import ExpressionAutoComplete from 'components/kit/ExpressionAutoComplete';
 import ErrorBoundary from 'components/ErrorBoundary/ErrorBoundary';
+import AutocompleteInput from 'components/AutocompleteInput';
 
 import { ANALYTICS_EVENT_KEYS } from 'config/analytics/analyticsKeysMap';
 
@@ -41,8 +41,8 @@ function SelectForm({
     };
   }, []);
 
-  function handleSearch(e: React.ChangeEvent<any>): void {
-    e.preventDefault();
+  function handleSearch(e?: React.ChangeEvent<any>): void {
+    e?.preventDefault();
     if (requestIsPending || searchButtonDisabled) {
       return;
     }
@@ -120,17 +120,12 @@ function SelectForm({
             >
               {selectedImagesData?.advancedMode ? (
                 <div className='SelectForm__textarea'>
-                  <ExpressionAutoComplete
-                    isTextArea={true}
-                    onExpressionChange={onSelectAdvancedQueryChange}
-                    onSubmit={handleSearch}
+                  <AutocompleteInput
+                    advanced
+                    context={selectFormData?.advancedSuggestions}
+                    onEnter={handleSearch}
+                    onChange={onSelectRunQueryChange}
                     value={selectedImagesData?.advancedQuery}
-                    placeholder='images.name in [“loss”, “accuracy”] and run.learning_rate > 10'
-                    options={[
-                      'images.name',
-                      'images.context',
-                      ...selectFormData.suggestions,
-                    ]}
                   />
                 </div>
               ) : (
@@ -247,12 +242,11 @@ function SelectForm({
           {selectedImagesData?.advancedMode ? null : (
             <ErrorBoundary>
               <div className='SelectForm__TextField'>
-                <ExpressionAutoComplete
-                  onExpressionChange={onSelectRunQueryChange}
-                  onSubmit={handleSearch}
-                  value={selectedImagesData?.query}
-                  options={selectFormData.suggestions}
-                  placeholder='Filter runs, e.g. run.learning_rate > 0.0001 and run.batch_size == 32'
+                <AutocompleteInput
+                  context={selectFormData?.suggestions}
+                  onEnter={handleSearch}
+                  onChange={onSelectRunQueryChange}
+                  value={selectedImagesData?.advancedQuery}
                 />
               </div>
             </ErrorBoundary>