+++ aim/web/ui/src/pages/RunDetail/RunDetailSettingsTab.tsx
@@ -84,7 +84,6 @@ function RunDetailSettingsTab({
             }}
           />
         </div>
-
         <ConfirmModal
           open={openDeleteModal}
           onCancel={handleDeleteModalClose}