+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Cython bindings for `PrefixView`, `TreeView`, `Container`, `ArrayView` (mahnerak)
 - Add ability to track and visualize stdout/stderr (mihran113, VkoHov)
 - Fix `AimLogger` deprecation issues related to release of PyTorch Lightning v1.5 (arnauddhaene)
+- Enable better autocomplete experience with monaco editor (arsengit)
 
 ### Fixes: